adds gameBoardScreenTextHelper.js
