+++ lib/gameBoard.js
@@ -1,6 +1,7 @@
-const EnemyPlane               = require('./enemyPlane.js');
-const UserMissile              = require('./userMissile.js');
-const GameBoardMissileHelpers  = require('./gameBoardMissileHelper.js');
+const EnemyPlane                 = require('./enemyPlane.js');
+const UserMissile                = require('./userMissile.js');
+const GameBoardMissileHelpers    = require('./gameBoardMissileHelper.js');
+const GameBoardScreenTextHelper  = require('./gameBoardScreenTextHelper.js');
 
 function GameBoard(canvas, ctx){
   this.startGame       = false;
@@ -24,6 +25,7 @@ function GameBoard(canvas, ctx){
   this.powerUps        = [];
   this.playerPower     = 1;
   this.missileHelpers  = new GameBoardMissileHelpers(this.userMissiles,this.enemyMissiles, this.shipX, this.shipY, this.canvasWidth, this.canvasHeight);
+  this.textHelpers     = new GameBoardScreenTextHelper(this.ctx, this.score, this.canvasWidth);
 }
 
 GameBoard.prototype.launchNewMissile = function(){
@@ -48,25 +50,13 @@ GameBoard.prototype.checkMissileLocation = function(){
     return (missile.x > 0 && missile.x < this.canvasWidth && missile.y > 0 && missile.y < this.canvasHeight);
   }.bind(this));
 };
-//
-// GameBoard.prototype.renderEnemyMissiles = function(){
-//   this.enemyMissiles.forEach(function(missile){
-//     missile.drawMissile().launch();
-//   });
-// };
-//
-// GameBoard.prototype.renderUserMissiles = function(){
-//   this.userMissiles.forEach(function(missile){
-//     missile.drawMissile().launch();
-//   });
-// };
 
 GameBoard.prototype.checkForSecondRound = function(){
   if (this.frameCount === 300){
     this.level = 2;
    }
    if(this.level === 2 && this.levelStart === false) {
-     this.levelScreen('2');
+     this.textHelpers.levelScreen('2');
      this.enemyPlanes = [];
    }
 
@@ -81,7 +71,7 @@ GameBoard.prototype.checkForThirdRound = function() {
     this.levelStart = false;
    }
    if(this.level === 3 && this.levelStart === false) {
-     this.levelScreen('3');
+     this.textHelpers.levelScreen('3');
      this.enemyPlanes = [];
    }
 
@@ -160,47 +150,10 @@ GameBoard.prototype.drawShip = function(){
   this.ctx.closePath();
 };
 
-GameBoard.prototype.levelScreen = function(levelNumber){
-  this.ctx.font = "20px 'Press Start 2P'";
-  this.ctx.fillStyle = "#0095DD";
-  let textString = "Starting Level " + levelNumber;
-  this.addTextToScreen(textString, 400);
-};
-
-GameBoard.prototype.drawScore = function(){
-  this.ctx.font = "16px 'Press Start 2P'";
-  this.ctx.fillStyle = "#0095DD";
-  this.ctx.fillText("Score: " + this.score, this.canvasWidth - 200, 20);
-};
-
-GameBoard.prototype.startScreen = function(){
-  this.ctx.font = "20px 'Press Start 2P'";
-  this.ctx.fillStyle = "#0095DD";
-  let textString = "Press Enter to Play";
-  this.addTextToScreen(textString, 400);
-};
-
-GameBoard.prototype.gameOverScreen = function(){
-  this.ctx.font = "20px 'Press Start 2P'";
-  this.ctx.fillStyle = "#0095DD";
-  let textString = "See You Space Cowboy...";
-  let textString2 = "End Score: " + this.score;
-  let textString3 = "Press P to Play Again";
-
-  this.addTextToScreen(textString, 300);
-  this.addTextToScreen(textString2, 400);
-  this.addTextToScreen(textString3, 500);
-};
-
-GameBoard.prototype.addTextToScreen = function(textString, y){
-  let textWidth = this.ctx.measureText(textString).width;
-  this.ctx.fillText(textString, (this.canvasWidth/2) - (textWidth/2),y);
-};
-
 GameBoard.prototype.setup = function() {
   this.checkForSecondRound();
   this.checkForThirdRound();
-  this.drawScore();
+  this.textHelpers.drawScore();
   this.powerUpDrop();
   this.missileHelpers.renderEnemyMissiles();
   this.renderEnemyPlanes();
