+++ lib/gameBoardMissileHelper.js
@@ -7,12 +7,6 @@ function GameBoardMissileHelpers(userMissiles,enemyMissiles, shipX, shipY, canva
   this.canvasHeight = canvasHeight;
 }
 
-// GameBoardMissileHelpers.prototype.checkMissileLocation = function(){
-//   this.userMissiles.filter(function (missile){
-//     return (missile.x > 0 && missile.x < this.canvasWidth && missile.y > 0 && missile.y < this.canvasHeight);
-//   }.bind(this));
-// };
-
 GameBoardMissileHelpers.prototype.renderEnemyMissiles = function(){
   this.enemyMissiles.forEach(function(missile){
     missile.drawMissile().launch();
