+++ lib/mainGameLoop.js
@@ -32,8 +32,8 @@ MainGameLoop.prototype.loop = function(){
   this.ctx.clearRect(0, 0, this.canvas.width, this.canvas.height);
   this.renderBackground();
 
-  if(this.gameBoard.startGame !== true){this.gameBoard.startScreen();}
-  if(this.gameBoard.endGame === true){this.gameBoard.gameOverScreen();}
+  if(this.gameBoard.startGame !== true){this.gameBoard.textHelpers.startScreen();}
+  if(this.gameBoard.endGame === true){this.gameBoard.textHelpers.gameOverScreen();}
   if(this.gameBoard.startGame && !this.gameBoard.endGame){
     this.gameBoard.setup();
     this.keyController.userMovement();
